fix: moved ssh module out of user
