+++ src/app.module.ts
@@ -10,7 +10,7 @@ import { APP_GUARD } from '@nestjs/core';
 import { JwtAuthGuard } from './auth/guards/jwt-auth.guard';
 import { AuthModule } from './auth/auth.module';
 import { RolesGuard } from './auth/guards/roles.guard';
-import { SshModule } from './users/ssh/ssh.module';
+import { SshModule } from './ssh/ssh.module';
 
 @Module({
   imports: [
